% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mb_modelmatrix}
\alias{mb_modelmatrix}
\title{Construct BART model matrix from a data frame or factor/numeric vector.}
\usage{
mb_modelmatrix(X)
}
\arguments{
\item{X}{}
}
\value{
A matrix with attributes indicating which columns are dummies constructed from a single factor
}
\description{
Adapted from bartModelMatrix in the BART package. This version only constructs
the design matrix, adds grouping information as attributes, and converts
character columns to factors.
}
